+++ packages/wire/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `provider_retry_wait_start` and `provider_retry_wait_end` to the collab `AgentEvent` union so a host can replicate a provider-internal retry backoff (and its countdown) to guests ([#12785](https://github.com/can1357/oh-my-pi/pull/12785) by [@geoyws](https://github.com/geoyws)).
+
 ## [18.2.5] - 2026-09-17
 
 ### Added